# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "知流 API",
    "version": "1.0.0",
    "description": "知流 Web 与 iOS 共用的普通用户 API 契约"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/chat/ask": {
      "post": {
        "operationId": "askKnowledgeBase",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskWithCredentialsInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/chat/conversations": {
      "get": {
        "operationId": "listConversations",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Conversation"
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createConversation",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/chat/conversations/{id}": {
      "get": {
        "operationId": "getConversation",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateConversation",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationPatchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteConversation",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/chat/conversations/{id}/messages/stream": {
      "post": {
        "operationId": "streamChatMessage",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 消息流",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/chat/messages/{id}/regenerate/stream": {
      "post": {
        "operationId": "regenerateChatMessage",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 消息流",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/chat/runs/{id}/cancel": {
      "post": {
        "operationId": "cancelChatRun",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/chat/messages/{id}/save-to-note": {
      "post": {
        "operationId": "saveChatMessageToNote",
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveMessageToNoteInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "请求无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ai_note_session"
      }
    },
    "schemas": {
...
      "AskWithCredentialsInput": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "scope": {
            "type": "string",
            "enum": [
              "NOTE",
              "FOLDER",
              "WORKSPACE"
            ],
            "default": "WORKSPACE"
          },
          "noteId": {
            "type": "string",
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "folderId": {
            "type": "string",
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "credentials": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "configId": {
                  "type": "string",
                  "pattern": "^[cC][^\\s-]{8,}$"
                },
                "apiKey": {
                  "type": "string",
                  "maxLength": 4096,
                  "default": ""
                }
              },
              "required": [
                "configId"
              ],
              "additionalProperties": false
            },
            "maxItems": 2,
            "default": []
          }
        },
        "required": [
          "question"
        ],
        "additionalProperties": false
      },
...
      "ConversationCreateInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "mode": {
            "type": "string",
            "enum": [
              "KNOWLEDGE",
              "GENERAL"
            ],
            "default": "KNOWLEDGE"
          },
          "scope": {
            "type": "string",
            "enum": [
              "NOTE",
              "FOLDER",
              "WORKSPACE"
            ],
            "default": "WORKSPACE"
          },
          "noteId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "folderId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "providerConfigId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          }
        },
        "additionalProperties": false
      },
      "ConversationPatchInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "archived": {
            "type": "boolean"
          },
          "mode": {
            "type": "string",
            "enum": [
              "KNOWLEDGE",
              "GENERAL"
            ]
          },
          "scope": {
            "type": "string",
            "enum": [
              "NOTE",
              "FOLDER",
              "WORKSPACE"
            ]
          },
          "noteId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "folderId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "providerConfigId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          }
        },
        "additionalProperties": false
      },
      "ChatMessageInput": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          },
          "providerConfigId": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^[cC][^\\s-]{8,}$"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "credentials": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "configId": {
                  "type": "string",
                  "pattern": "^[cC][^\\s-]{8,}$"
                },
                "apiKey": {
                  "type": "string",
                  "maxLength": 4096,
                  "default": ""
                }
              },
              "required": [
                "configId"
              ],
              "additionalProperties": false
            },
            "maxItems": 2,
            "default": []
          }
        },
        "required": [
          "content"
        ],
        "additionalProperties": false
      },
      "SaveMessageToNoteInput": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "mode": {
                "type": "string",
                "enum": [
                  "CREATE"
                ]
              },
              "title": {
                "type": "string",
                "minLength": 1,
                "maxLength": 200
              },
              "folderId": {
                "type": [
                  "string",
                  "null"
                ],
                "pattern": "^[cC][^\\s-]{8,}$"
              },
              "includeQuestion": {
                "type": "boolean",
                "default": true
              },
              "includeCitations": {
                "type": "boolean",
                "default": true
              }
            },
            "required": [
              "mode"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "mode": {
                "type": "string",
                "enum": [
                  "APPEND"
                ]
              },
              "noteId": {
                "type": "string",
                "pattern": "^[cC][^\\s-]{8,}$"
              },
              "version": {
                "type": "integer",
                "minimum": 0
              },
              "includeQuestion": {
                "type": "boolean",
                "default": true
              },
              "includeCitations": {
                "type": "boolean",
                "default": true
              }
            },
            "required": [
              "mode",
              "noteId",
              "version"
            ],
            "additionalProperties": false
          }
        ]
      },
...
      "JSONValue": {
        "description": "保留未知 Tiptap 节点和属性的任意 JSON 值"
      },
...
      "Note": {
        "type": "object",
        "required": [
          "id",
          "workspaceId",
          "title",
          "content",
          "plainText",
          "status",
          "aiEnabled",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/JSONValue"
          },
          "plainText": {
            "type": "string"
          },
          "folderId": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "INBOX",
              "ACTIVE",
              "ARCHIVED"
            ]
          },
          "aiEnabled": {
            "type": "boolean"
          },
          "version": {
            "type": "integer"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": true
      },
...
      "Conversation": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "Error": {
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/JSONValue"
          }
        }
      }
    }
  }
}